+++ src/main/java/com/mcrminer/model/projections/ProjectProjection.java
@@ -3,7 +3,7 @@ package com.mcrminer.model.projections;
 import java.util.Set;
 
 public interface ProjectProjection {
-    Integer getId();
+    Long getId();
     String getUrlPath();
     String getName();
     Set<ReviewRequestProjection> getReviewRequests();
