+++ src/main/java/com/mcrminer/repository/DiffRepository.java
@@ -3,5 +3,5 @@ package com.mcrminer.repository;
 import com.mcrminer.model.Diff;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DiffRepository extends JpaRepository<Diff, Integer> {
+public interface DiffRepository extends JpaRepository<Diff, Long> {
 }
